fix(commit): enforce fixed-bucket ownership globally
